requirements.txt
# Core Data Science
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# SQL & Data Engineering
duckdb>=0.10.0

# ML & Experimentation
xgboost>=2.0.0
mlflow>=2.10.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Dashboard
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
joblib>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.29.0
